Update hash_mash.rb
`.each` is only appropriate when the loop is being used for its side effect (writing to a file, e.g.), not for manipulating the elements of an array. 

This code properly belongs in the ActiveRecord models. Active users should be a scope on the user model, and the other method should be defined on the message model.